requirements.txt
pillow==12.0.0
psycopg2-binary==2.9.11
tk==0.1.0
pyinstaller==6.17.0